requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
enum34==1.1.6
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
peewee==3.8.2
pep8==1.7.1
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
pycodestyle==2.5.0
pylint==2.2.2
python-dateutil==2.7.5
requests==2.21.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.24.1
vk-api==11.3.0
wrapt==1.11.1
